refactor: relocate Gemini LLM initialization to summarize_other.py
Moved the Summary Pydantic model, page_of_ function, and Gemini LLM/chain
initialization logic from src/endpoint/readit/core.py to
src/endpoint/readit/app/summarize_other.py.

This change ensures that other tools (like fetch.py) that import
FetchResult from core.py do not trigger LLM initialization or require
the GEMINI_API_KEY environment variable. Initialization is now deferred
until it's actually needed within the main function of
summarize_other.py.
