+++ src/endpoint/readit/app/summarize_other.py
@@ -5,14 +5,36 @@ import os
 from urllib.parse import urlparse
 
 import click
+from langchain_google_genai import ChatGoogleGenerativeAI
+from langchain_core.prompts import ChatPromptTemplate
+from pydantic import BaseModel
+from pydantic import Field
 
 from endpoint.readit.core import Page
 from endpoint.readit.core import FetchResult
-from endpoint.readit.core import page_of_
 
 
 logger = getLogger(__name__)
 logger.setLevel(os.environ.get("ENTRYPOINT_LOG_LEVEL", "INFO").upper())
+
+
+class Summary(BaseModel):
+    title: str = Field(
+        description="The title of concise main title of the article or page"
+    )
+    date: str = Field(
+        description="The issue or publication date as YYYY/MM/DD format (????/??/?? if unknown)"
+    )
+
+
+def page_of_(fetch_result: FetchResult, chain) -> Page:
+    summary = chain.invoke({"content": fetch_result.html})
+
+    return Page(
+        url=urlparse(str(fetch_result.url)),
+        title=summary.title,
+        date=summary.date,
+    )
 
 
 @click.command()
@@ -44,7 +66,22 @@ def main(output_path: str, fetch_result_path: str) -> None:
             },
         )
     else:
-        page = page_of_(fetch_result)
+        prompt = ChatPromptTemplate.from_template("""
+            Analyze the following content from a webpage and extract two pieces of information:
+            1. The concise main title of the article or page.
+            2. The issue or publication date as YYYY/MM/DD format (if available).
+               - If not available, state "????/??/??".
+
+            Format your answer as a JSON object with keys "date" and "title".
+
+            Content: {content}
+            """)
+        structured_llm = ChatGoogleGenerativeAI(
+            model="gemini-2.5-flash"
+        ).with_structured_output(Summary)
+        chain = prompt | structured_llm
+
+        page = page_of_(fetch_result, chain)
 
     logger.info("Result: '%s'", page)
