+++ src/endpoint/readit/core.py
@@ -1,24 +1,12 @@
 from dataclasses import dataclass
 from dataclasses import field
 
-from langchain_google_genai import ChatGoogleGenerativeAI
-from langchain_core.prompts import ChatPromptTemplate
 from pydantic import BaseModel
-from pydantic import Field
 from pydantic import HttpUrl
 from typing import Any
 from urllib.parse import ParseResult as URL
 from urllib.parse import urlparse
 from urllib.parse import urlunparse
-
-
-class Summary(BaseModel):
-    title: str = Field(
-        description="The title of concise main title of the article or page"
-    )
-    date: str = Field(
-        description="The issue or publication date as YYYY/MM/DD format (????/??/?? if unknown)"
-    )
 
 
 class FetchResult(BaseModel):
@@ -56,29 +44,3 @@ class Page:
             kind=d.get("kind", "other"),
             metadata=d.get("metadata", {}),
         )
-
-
-_PROMPT = ChatPromptTemplate.from_template("""
-    Analyze the following content from a webpage and extract two pieces of information:
-    1. The concise main title of the article or page.
-    2. The issue or publication date as YYYY/MM/DD format (if available).
-       - If not available, state "????/??/??".
-
-    Format your answer as a JSON object with keys "date" and "title".
-
-    Content: {content}
-    """)
-_STRUCTURED_LLM = ChatGoogleGenerativeAI(
-    model="gemini-2.5-flash"
-).with_structured_output(Summary)
-_CHAIN = _PROMPT | _STRUCTURED_LLM
-
-
-def page_of_(fetch_result: FetchResult) -> Page:
-    summary = _CHAIN.invoke({"content": fetch_result.html})
-
-    return Page(
-        url=urlparse(str(fetch_result.url)),
-        title=summary.title,
-        date=summary.date,
-    )
